[IMP] event_sale: move registrationEditorLine to its file
This commit moves registrationEditorLine model into separate
file in order to have a consistency with all other models in
the wizard folder.
