+++ addons/event_sale/wizard/event_edit_registration_line.py
@@ -1,0 +1,33 @@
+# -*- coding: utf-8 -*-
+
+from odoo import models, fields
+
+class RegistrationEditorLine(models.TransientModel):
+    """Event Registration"""
+    _name = "registration.editor.line"
+    _description = 'Edit Attendee Line on Sales Confirmation'
+    _order = "id desc"
+
+    editor_id = fields.Many2one('registration.editor')
+    sale_order_line_id = fields.Many2one('sale.order.line', string='Sales Order Line')
+    event_id = fields.Many2one('event.event', string='Event', required=True)
+    registration_id = fields.Many2one('event.registration', 'Original Registration')
+    event_ticket_id = fields.Many2one('event.event.ticket', string='Event Ticket')
+    email = fields.Char(string='Email')
+    phone = fields.Char(string='Phone')
+    mobile = fields.Char(string='Mobile')
+    name = fields.Char(string='Name')
+
+    def get_registration_data(self):
+        self.ensure_one()
+        return {
+            'event_id': self.event_id.id,
+            'event_ticket_id': self.event_ticket_id.id,
+            'partner_id': self.editor_id.sale_order_id.partner_id.id,
+            'name': self.name or self.editor_id.sale_order_id.partner_id.name,
+            'phone': self.phone or self.editor_id.sale_order_id.partner_id.phone,
+            'mobile': self.mobile or self.editor_id.sale_order_id.partner_id.mobile,
+            'email': self.email or self.editor_id.sale_order_id.partner_id.email,
+            'sale_order_id': self.editor_id.sale_order_id.id,
+            'sale_order_line_id': self.sale_order_line_id.id,
+        }
